backend/requirements.txt
GitPython==3.1.43
langchain==0.2.1
langchain-community==0.2.1
langchain-openai==0.1.7
langchain-core==0.2.0
openai==1.35.3
tiktoken==0.7.0
uvicorn==0.23.2
fastapi==0.109.1
qdrant-client==1.9.0
python-dotenv==1.0.1
pydantic==1.10.13
orjson==3.9.15
PyMuPDF==1.23.6
redis==5.0.1
beautifulsoup4==4.12.2
truefoundry[ml]==0.2.4
markdownify==0.11.6
gunicorn==22.0.0
markdown-crawler==0.0.8
black==24.3.0
async-timeout==4.0.3
pre-commit==3.7.0
huggingface-hub==0.20.3
requests==2.32.2
python-multipart==0.0.9
opencv-python==4.9.0.80
pillow==10.3.0

### transformers
transformers==4.40.2

#SambaStudio
sseclient-py==1.8.0

### prisma
prisma==0.13.1

#Replicate
replicate==0.26.0

#LoggingTracing
langsmith==0.1.71

#unstrutured
unstructured[all-docs]==0.13.6
